#81 [REFACTOR] Auth 타입 분리 및 API 타입 구조 개선
